Update Handler Middleware and Error

+++ Backend/src/middleware/uploadHandler.js
@@ -1,3 +1,4 @@
+// Backend/middleware/uploadHandler.js
 import { upload } from '../config/multer.js';
 import cloudinary, { CLOUDINARY_FOLDERS } from '../config/cloudinary.js';
 import fs from 'fs';
@@ -8,6 +9,9 @@ export const uploadBlogThumbnail = upload.single('thumbnail');
 export const uploadBlogContentImages = upload.array('contentImages', 10);
 
 export const uploadToCloudinary = async (file, folder) =>{
+     if(!file || !file.path){
+          throw new Error("No valid file provided for upload");
+     }
      try {
           const result = await cloudinary.uploader.upload(file.path, {
                folder,
@@ -20,10 +24,12 @@ export const uploadToCloudinary = async (file, folder) =>{
                publicId: result.public_id,
           };
      } catch (error) {
-          fs.unlinkSync(file.path);
-          throw new Error('Cloudinary upload failed');
+          if(fs.unlinkSync(file.path)){
+               fs.unlinkSync(file.path);
+          }
+          throw new Error(`Cloudinary upload failed: ${error.message}`);
      }
-}
+};
 
 export const deleteFromCloudinary = async (publicId) => {
      try {
@@ -31,4 +37,4 @@ export const deleteFromCloudinary = async (publicId) => {
      } catch (error) {
          throw new Error('Cloudinary deletion failed'); 
      }
-}
+};
